Better delay handling in control.py
Now the arm delay doesn't stack with the regular instruction delay, and the expected runtime is very close to the real value (but we're still losing a few ms every loop).
Also added final runtime printout at the end, and some more detailed examples for the control sequences.
Added comment with the currently used stty settings for the Arduino Due.